backend/main_simple.py
"""
OncoPurpose Backend API - Simplified Launcher
"""

# Copyright (c) 2025 OncoPurpose (trovesx)
# All Rights Reserved.
# For licensing info, see LICENSE or contact oncopurpose@trovesx.com

import logging
import os
from pathlib import Path
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

# Load environment variables explicitly
from dotenv import load_dotenv
env_path = Path(__file__).parent / '.env'
load_dotenv(dotenv_path=env_path)

from config import settings
from discovery_api import router as discovery_router

# Setup logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Create FastAPI app
app = FastAPI(
    title="OncoPurpose API",
    description="Drug repurposing analysis platform",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc",
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=settings.allowed_origins_list,  # Use property instead
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Health check
@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "service": "OncoPurpose API",
        "version": "1.0.0"
    }

# Root endpoint
@app.get("/")
async def root():
    return {
        "message": "Welcome to OncoPurpose API",
        "docs": "/docs",
        "health": "/health",
        "discovery": "/api/discovery/analyze"
    }

# Include discovery router
app.include_router(discovery_router)

logger.info("OncoPurpose API started successfully")
logger.info(f"OpenAI API Key configured: {bool(settings.OPENAI_API_KEY)}")
logger.info(f"Using OpenAI model: {settings.OPENAI_MODEL}")

if __name__ == "__main__":
    import uvicorn
    
    uvicorn.run(
        "main_simple:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
