backend/requirement.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1
aioredis==2.0.1

# External API integrations
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
rdkit-pypi==2023.9.1

# Data processing
python-dateutil==2.8.2

# File handling
python-magic==0.4.27
boto3==1.34.0

# Background jobs
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8